Embeded Message Command
Here you can make an custom message so when you use the command in your Discord server, it will come up with the message you have instructed it. You could also copy this code to make an extra embeded message version. As many as you would like, but you have to change the name of it.